BUG: Fix display of screen value for greyscale images in new viewer

+++ Code/Visualization/otbStandardRenderingFunction.h
@@ -105,7 +105,7 @@ public:
   {
     itk::OStringStream oss;
     OutputPixelType output = this->Evaluate(spixel);
-    oss<<"Grayscale[value: "<< static_cast<typename itk::NumericTraits<ScalarPixelType>::PrintType>(spixel)<<", displayed: "<<output[0]<<"]";
+    oss<<"Grayscale[value: "<< static_cast<typename itk::NumericTraits<ScalarPixelType>::PrintType>(spixel)<<", displayed: "<< static_cast<unsigned int>(output[0])<<"]";
     return oss.str();
   }
